data_science_2/requirements.txt
contourpy==1.3.2
cycler==0.12.1
dotenv==0.9.9
fonttools==4.58.4
joblib==1.5.1
kiwisolver==1.4.8
matplotlib==3.10.3
mysql-connector-python==9.3.0
numpy==2.2.6
packaging==25.0
pandas==2.3.0
pillow==11.3.0
psycopg2-binary==2.9.10
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
